requirements.txt
# HighEfficiencyGlide - Dependencies
# Solar-powered autonomous glider simulation and control

# Core simulation
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
pyyaml>=6.0

# 3D visualization
pyvista>=0.38.0
vtk>=9.2.0

# ArduPilot/MAVLink
pymavlink>=2.4.0
dronekit>=2.9.2
MAVProxy>=1.8.60

# YOLO/Vision
ultralytics>=8.0.0
opencv-python>=4.7.0
torch>=2.0.0
torchvision>=0.15.0

# Model optimization (for Jetson deployment)
onnx>=1.14.0
onnxruntime>=1.15.0

# Ground station
fastapi>=0.100.0
websockets>=11.0
uvicorn>=0.22.0
jinja2>=3.1.0

# Data handling
pandas>=2.0.0
h5py>=3.8.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
